latediff

+++ controllers/ManagingUser.py
@@ -21,7 +21,15 @@ class ManagingUser:
     def CreateUser(self, user_item: UserDB):
         # Use admin password from .env for DB operations (dummy for SQLite)
         admin_password = os.getenv("ADMIN_PASSWORD", "admin123")
+        loggedUser = self.LoginUser(None, None)
+        if loggedUser is None or loggedUser.role.lower()!="Gestion":
+            print("Authentification échouée.")
+            return None
         engine = self.db.LoginDatabase(email="admin@epicevents.com", password=admin_password)
+        if engine is None:
+            print("Connection échouée.")
+            return None
+        
         with Session(engine) as session:
             session.add(user_item)
             session.commit()
@@ -34,8 +42,17 @@ class ManagingUser:
         return False
 
     def DeleteUser(self, user_id: int):
+        # Use admin password from .env for DB operations (dummy for SQLite)
         admin_password = os.getenv("ADMIN_PASSWORD", "admin123")
+        loggedUser = self.LoginUser(None, None)
+        if loggedUser is None or loggedUser.role.lower()!="Gestion":
+            print("Authentification échouée.")
+            return None
         engine = self.db.LoginDatabase(email="admin@epicevents.com", password=admin_password)
+        if engine is None:
+            print("Connection échouée.")
+            return None
+        
         with Session(engine) as session:
             usr = session.get(UserDB, user_id)
             session.delete(usr)
@@ -55,9 +72,9 @@ class ManagingUser:
             email = input()
             print("Login password: ")
             user_password = input()
-            salt = bcrypt.gensalt()
-            hashed_password = bcrypt.hashpw(user_password.encode('utf-8'), salt).decode('utf-8')
-            password_hash = hashed_password
+            #salt = bcrypt.gensalt()
+            #hashed_password = bcrypt.hashpw(user_password.encode('utf-8'), salt).decode('utf-8')
+            #password_hash = hashed_password
 
         with Session(engine) as session:
             stmt = select(UserDB).where(UserDB.email == email)
@@ -67,7 +84,7 @@ class ManagingUser:
             if not user:
                 return None
                         
-            if bcrypt.checkpw(password_hash.encode('utf-8'), user.password.encode('utf-8')):
+            if bcrypt.checkpw(user_password.encode('utf-8'), user.password.encode('utf-8')):
                 print("Correct password")
                 self.db.currentUser = user
                 return user
